requirements.txt
openai-whisper
pytesseract
opencv-python
Pillow
anthropic
